requirements.txt
Flask~=1.1.2
Flask-Script~=2.0.6
Flask-SQLAlchemy~=2.4.4
flask_marshmallow~=0.13.0
marshmallow-sqlalchemy~=0.23.1
flask-restx~=0.2.0
python-dotenv~=0.14.0
gunicorn~=20.0.4
click~=7.1.2
SQLAlchemy~=1.3.18
alembic~=1.4.2